COMP:Fixed old-style cast warnings.

+++ VTK/Geovis/vtkGeoCamera.cxx
@@ -269,8 +269,8 @@ void vtkGeoCamera::InitializeNodeAnalysis(int rendererSize[2])
 {
   this->Aspect[1] = tan(this->VTKCamera->GetViewAngle() 
                           * vtkMath::DegreesToRadians() * 0.5);
-  this->Aspect[0] = this->Aspect[1] * (double)(rendererSize[0]) 
-                       / (double)(rendererSize[1]);
+  this->Aspect[0] = this->Aspect[1] * rendererSize[0]
+                       / static_cast<double>(rendererSize[1]);
 
   this->VTKCamera->GetViewPlaneNormal(this->ForwardNormal);
   this->ForwardNormal[0] = - this->ForwardNormal[0];
